backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "orjson>=3.10.18",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
